requirements.txt
Appium-Python-Client==0.24
attrs==17.2.0
colorama==0.3.9
colormap==1.0.1
easydev==0.9.35
enum34==1.1.6
et-xmlfile==1.0.1
jdcal==1.3
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.3
openpyxl==2.4.9
pandas==0.21.0
pexpect==4.2.1
pluggy==0.5.2
ptyprocess==0.5.2
py==1.4.34
python-dateutil==2.6.1
pytz==2017.3
reflection==0.1.1
reports==0.3.1
selenium==3.7.0
six==1.11.0
